+++ exporter/jaegerexporter/factory_test.go
@@ -50,7 +50,7 @@ func TestCreateInstanceViaFactory(t *testing.T) {
 	// Default config doesn't have default endpoint so creating from it should
 	// fail.
 	params := component.ExporterCreateParams{Logger: zap.NewNop()}
-	exp, err := factory.CreateTraceExporter(context.Background(), params, cfg)
+	exp, err := factory.CreateTracesExporter(context.Background(), params, cfg)
 	assert.NotNil(t, err)
 	assert.Equal(t, "\"jaeger\" config requires a non-empty \"endpoint\"", err.Error())
 	assert.Nil(t, exp)
@@ -58,7 +58,7 @@ func TestCreateInstanceViaFactory(t *testing.T) {
 	// Endpoint doesn't have a default value so set it directly.
 	expCfg := cfg.(*Config)
 	expCfg.Endpoint = "some.target.org:12345"
-	exp, err = factory.CreateTraceExporter(context.Background(), params, cfg)
+	exp, err = factory.CreateTracesExporter(context.Background(), params, cfg)
 	assert.NoError(t, err)
 	assert.NotNil(t, exp)