+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - `kafka` exporter: Add support for exporting metrics as otlp Protobuf.
 - Rename component.TraceReceiver to component.TracesReceiver #1975
 - Rename component.TraceProcessor to component.TracesProcessor #1976
+- Rename component.TraceExporter to component.TracesExporter #1975
 
 ## v0.13.0 Beta